ekyc/requirements.txt
# Core Web Framework
fastapi==0.116.1
uvicorn==0.35.0
aiofiles==24.1.0
python-multipart==0.0.20
streamlit>=1.24.0

# AI and LLM
openai==1.97.1
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.28
langchain-elasticsearch==0.3.2
sentence-transformers==5.0.0
transformers==4.54.0
torch>=2.0.0
tiktoken==0.9.0

# Vector Database
elasticsearch==8.18.1
faiss-cpu==1.11.0.post1

# Data Processing
numpy>=1.25.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Environment and Configuration
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1

# HTTP and async
httpx==0.28.1
aiohttp==3.12.14
requests>=2.31.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.65.0
Pillow>=10.0.0
PyYAML>=6.0.1

# OCR and Image Processing
pytesseract>=0.3.10
easyocr>=1.7.0
opencv-python>=4.8.0
pdf2image>=1.16.3